+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ History
 Version 2.2.2 (in development)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-- Sync with upstream pathlib from CPython 3.6.1 (bpo-29694).
+- Sync with upstream pathlib from CPython 3.6.1 (bpo-25803, bpo-29694).
 
 Version 2.2.1
 ^^^^^^^^^^^^^